requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
torchvision
Pillow
numpy
pydantic_settings
boto3